feature(signin-database) Take login data from the database  [starts #166150980]
