+++ server/controller/loginController.js
@@ -1,40 +1,39 @@
-import { getSingleUser } from '../helper/userHelper';
-
-const wrongDataStatus = (res, message) => {
-  res.status(400).send({
-    status: 400,
-    message,
-  });
-};
+import { comparePassword } from '../helper/hashPassword';
+import { getSigninQuery } from '../models/Queries';
+import pool from '../config/configDb';
 
 export default function login(req, res) {
   const { email, password } = req.body;
-  if (email === undefined) {
-    res.status(400).send({
-      status: 400,
-      message: 'The email is required',
-    });
-  } else if (password === undefined) {
-    wrongDataStatus(res, 'The password is required');
-  } else if (!password || !email) {
-    res.status(401).send({
-      status: 401,
-      message: 'You provided a wrong email or password',
-    });
-  } else {
-    const [user] = getSingleUser(email);
 
-    if (user && user.validatePassword(password)) {
-      return res.status(200).json({
-        status: 200,
-        data: user.toJSON(),
+  const userDataArray = [email];
+  pool.query(getSigninQuery(userDataArray))
+    .then((result) => {
+      if (result.rowCount > 0) {
+        const user = result.rows[0];
+        if (user && comparePassword(password, user.password)) {
+          user.password = '';
+          res.status(200).json({
+            status: 200,
+            data: user,
+          });
+        } else {
+          res.status(401).send({
+            status: 401,
+            message: 'You provided a wrong email or password',
+          });
+        }
+      } else {
+        res.status(401).send({
+          status: 401,
+          message: 'You provided a wrong email or password',
+        });
+      }
+    })
+    .catch((err) => {
+      res.status(401).send({
+        status: 401,
+        message: 'You provided a wrong email or password',
       });
-    }
-
-    return res.status(401).send({
-      status: 401,
-      message: 'You provided a wrong email or password',
     });
-  }
   return res;
 }
